Fetch tab entries from server and persist them in room

+++ features/tabgroup/database/src/commonMain/kotlin/de/tabmates/features/tabgroup/database/entities/TabEntrySplitEntity.kt
@@ -32,6 +32,6 @@ data class TabEntrySplitEntity(
     val tabEntryId: String,
     val participantId: String,
     val splitType: SplitTypeDatabase,
-    val value: Float,
-    val resolvedAmount: Float,
+    val value: Double,
+    val resolvedAmount: Double,
 )
